+++ app/src/main/AndroidManifest.xml
@@ -23,19 +23,19 @@
         android:theme="@style/Theme.Cozymate"
         tools:targetApi="31">
         <activity
-            android:name=".ui.my_page.UpdateRoomInfoActivity"
+            android:name=".ui.my_page.update_room.UpdateRoomInfoActivity"
             android:exported="false" />
         <activity
             android:name=".ui.notification.NotificationActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.my_page.FetchLifestyleActivity"
+            android:name=".ui.my_page.lifestyle.FetchLifestyleActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.my_page.MyFavoriteActivity"
+            android:name=".ui.my_page.favorite.MyFavoriteActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.my_page.UpdateMyInfoActivity"
+            android:name=".ui.my_page.my_info.UpdateMyInfoActivity"
             android:exported="false" />
         <activity
             android:name=".ui.cozy_home.roommate.search_roommate.SearchRoommateActivity"
@@ -47,7 +47,7 @@
             android:name=".ui.cozy_home.room_detail.RoomDetailActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.cozy_home.room_detail.UpdateMyRoomInfoActivity"
+            android:name=".ui.my_page.update_room.UpdateMyRoomInfoActivity"
             android:exported="false" />
         <activity
             android:name=".ui.cozy_home.roommate.roommate_detail.CozyHomeRoommateDetailActivity"
@@ -117,9 +117,9 @@
         <activity android:name=".ui.message.MessageDetailActivity" />
         <activity android:name=".ui.message.MessageMemberActivity" />
         <activity android:name=".ui.message.WriteMessageActivity" />
-        <activity android:name=".ui.my_page.WithDrawActivity" />
-        <activity android:name=".ui.my_page.InquiryActivity" />
-        <activity android:name=".ui.my_page.WriteInquiryActivity" />
+        <activity android:name=".ui.my_page.withdraw.WithDrawActivity" />
+        <activity android:name=".ui.my_page.inquiry.InquiryActivity" />
+        <activity android:name=".ui.my_page.inquiry.WriteInquiryActivity" />
         <activity
             android:name=".ui.role_rule.AddTodoActivity"
             android:windowSoftInputMode="adjustNothing" />
