+++ app/src/main/java/umc/cozymate/ui/my_page/inquiry/WriteInquiryActivity.kt
@@ -1,4 +1,4 @@
-package umc.cozymate.ui.my_page
+package umc.cozymate.ui.my_page.inquiry
 
 import android.content.Intent
 import android.graphics.Color
