+++ app/src/main/java/umc/cozymate/ui/my_page/my_info/UpdateMyInfoActivity.kt
@@ -1,4 +1,4 @@
-package umc.cozymate.ui.my_page
+package umc.cozymate.ui.my_page.my_info
 
 import android.graphics.Color
 import android.os.Bundle
