+++ app/src/main/java/umc/cozymate/ui/my_page/my_info/UpdateNicknameFragment.kt
@@ -1,4 +1,4 @@
-package umc.cozymate.ui.my_page
+package umc.cozymate.ui.my_page.my_info
 
 import android.os.Bundle
 import android.text.Editable
