+++ app/src/main/java/umc/cozymate/ui/my_page/update_room/UpdateRoomInfoActivity.kt
@@ -1,8 +1,7 @@
-package umc.cozymate.ui.my_page
+package umc.cozymate.ui.my_page.update_room
 
 import android.graphics.Color
 import android.os.Bundle
-import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
