+++ app/src/main/res/layout/activity_my_favorite.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.my_page.MyFavoriteActivity">
+    tools:context=".ui.my_page.favorite.MyFavoriteActivity">
 
     <ImageView
         android:id="@+id/iv_back"
@@ -58,6 +58,7 @@
         android:layout_height="wrap_content"
         android:layout_marginHorizontal="20dp"
         android:layout_marginTop="20dp"
+        android:layout_marginBottom="40dp"
         android:clipChildren="false"
         android:visibility="gone"
         app:layout_constraintEnd_toEndOf="parent"
