+++ app/src/main/res/layout/activity_update_my_info.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.my_page.UpdateMyInfoActivity">
+    tools:context=".ui.my_page.my_info.UpdateMyInfoActivity">
 
     <FrameLayout
         android:id="@+id/fragment_update_my_info"
