+++ app/src/main/res/layout/activity_update_room_info.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.my_page.UpdateRoomInfoActivity">
+    tools:context=".ui.my_page.update_room.UpdateRoomInfoActivity">
 
     <FrameLayout
         android:id="@+id/fragment_update_room_info"
